+++ changelog.md
@@ -1,31 +1,49 @@
 **CHANGELOG**
 
 ***v0.1.2***
+
 Date: 2017-07-12
+
 Release type: Initial release
 ---
 * Initial release
 
 ***v0.1.3***
+
 Date: 2017-07-13
+
 Release type: Cleanup / Non-critical
 ---
 * Fixed an issue with the readme
 * Code clean-up
-* Added further Dev-Dependencies (tslint, typedoc)
+* Added further dev-dependencies (tslint, typedoc)
 
 ***v0.1.4***
+
 Date: 2017-07-17
+
 Release type: Bugfix
 ---
 * Fixed a bug regarding comparison
 * Code clean-up
+* Added dependency lodash.isequal
 
 ***v0.1.5***
+
 Date: 2017-07-17
+
 Release type: Bugfix
 ---
 * Fixed a bug in removeAt
 
 
+***v0.1.6***
+
+Date: 2017-07-17
+
+Release type: Bugfix
+---
+* Several bugfixes
+* Optimization of the List Class
+* Change of dependency from lodash to lodash.isequal